rename package from vault to hcvault
